money_maker_agent/requirements.txt
# Money Maker Agent Requirements

# Core dependencies
asyncio-compat>=0.1.0
aiohttp>=3.9.0
pyyaml>=6.0.1
watchdog>=3.0.0

# Database
# SQLite is included in Python standard library

# Web scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0
scrapy>=2.11.0

# Cryptography for account encryption
cryptography>=41.0.0

# Optional: Plaid integration for bank accounts
# plaid-python>=9.0.0

# Optional: LLM APIs (for content generation)
# openai>=1.0.0
# google-generativeai>=0.3.0

# Optional: Investment APIs (for portfolio agent)
# alpaca-trade-api>=3.0.0
# yfinance>=0.2.0

# Optional: WordPress API
# python-wordpress-xmlrpc>=2.3

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

